backend/requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
autopep8==2.0.0
az-iranian-bank-gateways==1.8.3
beautifulsoup4==4.11.1
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==38.0.4
daphne==4.0.0
Django==4.1.3
django-bootstrap-datepicker-plus==5.0.2
django-bootstrap-v5==1.0.11
django-cors-headers==3.13.0
django-extensions==3.2.1
django-jalali==6.0.0
django-jalali-date==1.0.1
django-leaflet==0.28.3
django-webpush==0.3.4
djangorestframework==3.14.0
http-ece==1.1.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
isodate==0.6.1
jdatetime==4.1.0
lxml==4.9.2
msgpack==1.0.4
Pillow==9.3.0
platformdirs==2.6.2
psycopg2-binary==2.9.5
py-vapid==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pycryptodome==3.16.0
pydantic==1.10.2
pygraphviz==1.10
pyOpenSSL==22.1.0
pytz==2022.6
pywebpush==1.9.4
redis==4.4.0
requests==2.28.1
requests-file==1.5.1
requests-toolbelt==0.10.1
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tomli==2.0.1
Twisted==22.10.0
txaio==22.2.1
typing_extensions==4.4.0
urllib3==1.26.13
zeep==4.2.1
zope.interface==5.5.2
